Split files and export edges
