+++ src/stories/DummyData.ts
@@ -75,73 +75,6 @@ export const nodes1 = [
 	}
 ]
 
-export const edges1 = [
-	{
-		id: 'e12',
-		source: '1',
-		target: '2',
-		type: 'smart',
-		markerEnd: { type: markerEndType }
-	},
-	{
-		id: 'e13',
-		source: '1',
-		target: '3',
-		type: 'smart',
-		markerEnd: { type: markerEndType }
-	},
-	{
-		id: 'e22a',
-		source: '2',
-		target: '2a',
-		type: 'smart',
-		markerEnd: { type: markerEndType }
-	},
-	{
-		id: 'e22b',
-		source: '2',
-		target: '2b',
-		type: 'smart',
-		markerEnd: { type: markerEndType }
-	},
-	{
-		id: 'e22c',
-		source: '2',
-		target: '2c',
-		type: 'smart',
-		markerEnd: { type: markerEndType }
-	},
-	{
-		id: 'e2c2d',
-		source: '2c',
-		target: '2d',
-		type: 'smart',
-		markerEnd: { type: markerEndType }
-	},
-	{
-		id: 'e2d2c',
-		source: '2d',
-		target: '2c',
-		type: 'smart',
-		markerEnd: { type: markerEndType }
-	},
-	{
-		id: 'e2d1',
-		source: '2d',
-		target: '1',
-		type: 'smart',
-		markerEnd: { type: markerEndType },
-		label: 'Node 2d to Node 1'
-	},
-	{
-		id: 'e2a2a',
-		source: '2a',
-		target: '2a',
-		type: 'smart',
-		markerEnd: { type: markerEndType }
-	}
-]
-
 export const nodes2 = [
 	{
 		id: '1',
@@ -155,12 +88,89 @@ export const nodes2 = [
 	}
 ]
 
-export const edges2 = [
+export const edges1Bezier = [
+	{
+		id: 'e12',
+		source: '1',
+		target: '2',
+		type: 'smartBezier',
+		markerEnd: { type: markerEndType }
+	},
+	{
+		id: 'e13',
+		source: '1',
+		target: '3',
+		type: 'smartBezier',
+		markerEnd: { type: markerEndType }
+	},
+	{
+		id: 'e22a',
+		source: '2',
+		target: '2a',
+		type: 'smartBezier',
+		markerEnd: { type: markerEndType }
+	},
+	{
+		id: 'e22b',
+		source: '2',
+		target: '2b',
+		type: 'smartBezier',
+		markerEnd: { type: markerEndType }
+	},
+	{
+		id: 'e22c',
+		source: '2',
+		target: '2c',
+		type: 'smartBezier',
+		markerEnd: { type: markerEndType }
+	},
+	{
+		id: 'e2c2d',
+		source: '2c',
+		target: '2d',
+		type: 'smartBezier',
+		markerEnd: { type: markerEndType }
+	},
+	{
+		id: 'e2d2c',
+		source: '2d',
+		target: '2c',
+		type: 'smartBezier',
+		markerEnd: { type: markerEndType }
+	},
+	{
+		id: 'e2d1',
+		source: '2d',
+		target: '1',
+		type: 'smartBezier',
+		markerEnd: { type: markerEndType },
+		label: 'Node 2d to Node 1'
+	},
+	{
+		id: 'e2a2a',
+		source: '2a',
+		target: '2a',
+		type: 'smartBezier',
+		markerEnd: { type: markerEndType }
+	}
+]
+
+export const edges2Bezier = [
 	{
 		id: 'e21',
 		source: '2',
 		target: '1',
-		type: 'smart',
+		type: 'smartBezier',
 		label: 'Label'
 	}
 ]
+
+export const edges1Straight = edges1Bezier.map((edge) => ({
+	...edge,
+	type: 'smartStraight'
+}))
+
+export const edges2Straight = edges2Bezier.map((edge) => ({
+	...edge,
+	type: 'smartStraight'
+}))
